+++ src/App.tsx
@@ -1,10 +1,15 @@
 import React from 'react';
 import './App.scss';
-import {Navigation} from './Router';
+import {Navigation} from './TempNav';
+import {BrowserRouter as Router, Route, Switch} from "react-router-dom";
+import {Curiosity} from "./Pages/Curiosity/CuriosityLandingPage";
+import {Spirit} from "./Pages/Spirit/SpiritLandingPage";
+import {Opportunity} from "./Pages/Opportunity/OpportunityLandingPage";
 
 
 function App() {
   return (
+      <Router>
     <div>
       <header className="header">
         <nav>
@@ -12,7 +17,19 @@ function App() {
           <Navigation />
         </nav>
       </header>
+        <Switch>
+            <Route exact path="/curiosity">
+                <Curiosity />
+            </Route>
+            <Route path="/spirit">
+                <Spirit />
+            </Route>
+            <Route path="/opportunity">
+                <Opportunity />
+            </Route>
+        </Switch>
     </div>
+      </Router>
   );
 }
